pyproject.toml
[project]
name = "bootcamp-aberto-ao-vivo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-airbyte>=5.2.1",
    "apache-airflow-providers-dbt-cloud>=4.4.1",
]
